+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 * Bugfixes
 * `image_to_base64` util to convert PIL -> base64 strings
+* UMAP tools: `upyog.ml.umap`
+* 
 
 
 ## 0.7.7 -- 28 Aug 2023